Refactored PI-specific invoices
The PI-specific is now implemented in the `PIInvoice` class.
This class takes in the entire billable invoice and will
export the PI invoices to local storage and S3

Because the `upload_to_s3` function is no longer used, it is removed.
Since the `export_s3` function in the Invoice class does not need to be
concerned about file extensions, the test case for s3 exporting is
somewhat simplified, only checking that the format of the output paths
are correct.