Restructured institute_map.json into a YAML list
As prerequisite to adding metadata about our list of institutions,
such as whether MGHPCC has a partnership with them, and to make the
file more human-friendly, `institute.json` has been converted to a YAML
and formatted as a list. Each element of the YAML list must be a dict with
2 attributes:
- `display_name`: The name of the institute, as will appear on invoices
- `domains`: A list containing domains for that institute

Additional metadata can be freely added to each institute dict as our
billing needs change.

There is also some small cleanup. Namely, some functions in `process_report.py`
that have been moved to `util.py` were not removed during refactoring.