Exported PI invoices as PDFs
The PI-specific dataframes will first be converted to HTML
tables using Jinja templates, and then converted to PDFs using
Chromium. Now, users of the script must provide a path to the
Chromium/Chrome binary throught the env var `CHROME_BIN_PATH`

A html template folder has been added, and the test cases
for the PI-specific invoice will now both check whether the
dataframe is formatted correctly and if the PDFs are
correctly generated. The dockerfile has been to install chromium