Correctly convert nerc-rates value to boolean
Closes #130

All values in nerc-rates are of type string and
need to be correctly interpreted by the caller.

Value for Limit New PI Credit to MGHPCC Partners is of type
string and value "True" and is being passed as such to a method
expecting a boolean.

This change add a comparison with the string "True" to perform
the conversion.